webmethods-to-boomi/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
duckduckgo-search==7.1.0
